pyproject.toml
[tool.poetry]
name = "BaseApp"
version = "0.0.1"
description = "Base FastAPI app"
authors = ["Kuvaev Maksim <shlyabom@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
grpcio = "^1.49.1"
grpcio-tools = "^1.49.1"
googleapis-common-protos = "^1.56.4"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
pytest = "^7.1.2"
flake8-bugbear = "^22.3.23"
flake8-comprehensions = "^3.8.0"
flake8-docstrings = "^1.6.0"
flake8-eradicate = "^1.2.0"
flake8-isort = "^4.1.1"
pep8-naming = "^0.12.1"
flake8-string-format = "^0.3.0"
flake8-bandit = "^3.0.0"
flake8-builtins = "^1.5.3"
flake8-black = "^0.3.2"
flake8-expression-complexity = "^0.0.11"
flake8-annotations-complexity = "^0.0.7"
flake8-requirements = "^1.5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
